tests/wpt/tests/css/filter-effects/css-filters-opacity-hit-testing.html

<!DOCTYPE html>
<title>Elements with 'opacity: 0' should respond to hit testing.</title>
<link rel="author" title="Psychpsyo" href="psychpsyo@gmail.com">
<link rel="help" href="https://drafts.fxtf.org/filter-effects/#FilterProperty">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  div {
    width: 100px;
    height: 100px;
    opacity: 0;
  }
</style>
<div></div>
<script>
  test(() => {
    assert_equals(document.elementFromPoint(50, 50).tagName, "DIV", "element with 'opacity: 0' doesn't respond to hit testing");
  });
</script>
